+++ doc/scripting/custom-actions.qml
@@ -1,5 +1,5 @@
 import QtQml 2.0
-import com.qownnotes.noteapi 1.0
+import QOwnNotesTypes 1.0
 
 /**
  * This script is an example for creating custom actions, that can be invoked