+++ doc/scripting/meeting-note.qml
@@ -1,5 +1,5 @@
 import QtQml 2.0
-import com.qownnotes.noteapi 1.0
+import QOwnNotesTypes 1.0
 
 /**
  * This script creates a menu item and a button to create or jump to the current date's meeting note