+++ doc/scripting/note-stats.qml
@@ -1,5 +1,5 @@
 import QtQml 2.0
-import com.qownnotes.noteapi 1.0
+import QOwnNotesTypes 1.0
 
 /*  This script shows current note statistics in a "label":
     Char(+s) = characters including spaces