+++ doc/scripting/paste-latex-image.qml
@@ -1,5 +1,5 @@
 import QtQml 2.0
-import com.qownnotes.noteapi 1.0
+import QOwnNotesTypes 1.0
 
 /**
  * This script creates a menu item and a button that generates and pastes an image from