+++ src/entities/script.cpp
@@ -1,3 +1,17 @@
+/*
+ * Copyright (c) 2014-2017 Patrizio Bekerle -- http://www.bekerle.com
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; version 2 of the License.
+ *
+ * This program is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
+ * or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License
+ * for more details.
+ *
+ */
+
 #include "script.h"
 #include <QDebug>
 #include <QSqlDatabase>
@@ -32,6 +46,16 @@ int Script::getId() {
  */
 QString Script::getScriptPath() {
     return this->scriptPath;
+}
+
+/**
+ * Returns the directory of the script
+ *
+ * @return
+ */
+QString Script::getScriptDirPath() {
+    QFileInfo info(scriptPath);
+    return info.canonicalPath();
 }
 
 QString Script::getName() {