+++ src/entities/script.h
@@ -1,3 +1,17 @@
+/*
+ * Copyright (c) 2014-2017 Patrizio Bekerle -- http://www.bekerle.com
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; version 2 of the License.
+ *
+ * This program is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
+ * or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License
+ * for more details.
+ *
+ */
+
 #pragma once
 
 #include <QSqlQuery>
@@ -5,6 +19,7 @@
 #include <QJsonObject>
 #include <QUrl>
 #include <QStringList>
+#include <QtCore/QDir>
 
 class Script
 {
@@ -46,6 +61,7 @@ public:
     void setSettingsVariablesJson(QString json);
     QJsonObject getSettingsVariablesJsonObject();
     void setSettingsVariablesJson(QJsonObject jsonObject);
+    QString getScriptDirPath();
 
 private:
     int id;